% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_filepaths}
\alias{add_filepaths}
\title{Add file paths to col_data}
\usage{
add_filepaths(
  col_data,
  results_dir = "/nfcore-rnaseq-pipeline/results_grcm38/star_salmon/",
  quant_file = "quant.sf"
)
}
\arguments{
\item{col_data}{Column data tibble from get_coldata()}

\item{results_dir}{Directory that directly contains sample subfolders
(e.g. "/nfcore-rnaseq-pipeline/results_grcm38/star_salmon/")}

\item{quant_file}{Quantification file name (default: "quant.sf")}
}
\value{
col_data with filepath column added
}
\description{
Add file paths to col_data
}
