% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_deseq_dataset}
\alias{create_deseq_dataset}
\title{Create DESeq2 dataset from tximport object}
\usage{
create_deseq_dataset(
  txi,
  col_data,
  level = "transcript",
  design = ~source + treatment + source:treatment,
  verbose = TRUE
)
}
\arguments{
\item{txi}{tximport object}

\item{col_data}{Column data for samples}

\item{level}{Quantification level used ("transcript" or "gene")}

\item{design}{Design formula (default: ~ source + treatment + source:treatment)}

\item{verbose}{Print dimensions and size factors (default: TRUE)}
}
\value{
DESeqDataSet object after running DESeq()
}
\description{
Create DESeq2 dataset from tximport object
}
