% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_tx2gene}
\alias{create_tx2gene}
\title{Create transcript-to-gene mapping from GTF}
\usage{
create_tx2gene(gtf, verbose = TRUE)
}
\arguments{
\item{gtf}{GTF object from rtracklayer::import()}

\item{verbose}{Print number of mappings (default: TRUE)}
}
\value{
data.frame with tx and gene columns
}
\description{
Create transcript-to-gene mapping from GTF
}
