% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_gtf}
\alias{load_gtf}
\title{Load GTF annotation file}
\usage{
load_gtf(gtf_path)
}
\arguments{
\item{gtf_path}{Path to GTF file}
}
\value{
GTF object from rtracklayer
}
\description{
Load GTF annotation file
}
