% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{define_contrasts}
\alias{define_contrasts}
\title{Define contrasts for differential expression analysis}
\usage{
define_contrasts(factor_levels, verbose = TRUE)
}
\arguments{
\item{factor_levels}{List with treatment_a, treatment_b, source_a, source_b}

\item{verbose}{Print contrast info (default: TRUE)}
}
\value{
Tibble with contrast definitions
}
\description{
Define contrasts for differential expression analysis
}
