% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.run_single_level_analysis}
\alias{.run_single_level_analysis}
\title{Helper function to run single-level analysis}
\usage{
.run_single_level_analysis(
  col_data,
  tx2gene,
  level,
  contrasts_info,
  design,
  alpha,
  verbose
)
}
\description{
Helper function to run single-level analysis
}
\keyword{internal}
